+++ support-frontend/assets/pages/[countryGroupId]/guardianAdLiteLanding/components/guardianAdLiteIconLeftSvg.tsx
