+++ support-frontend/assets/pages/[countryGroupId]/guardianAdLiteLanding/components/guardianAdLiteIconRightSvg.tsx
