+++ support-frontend/assets/pages/[countryGroupId]/router.tsx
@@ -4,7 +4,7 @@ import { setUpTrackingAndConsents } from 'helpers/page/page';
 import { renderPage } from 'helpers/rendering/render';
 import { geoIds } from 'pages/geoIdConfig';
 import { Checkout } from './checkout';
-import { GuardianLightLanding } from './guardianLightLanding/guardianLightLanding';
+import { GuardianLightLanding } from './guardianAdLiteLanding/guardianLightLanding';
 import { OneTimeCheckout } from './oneTimeCheckout';
 import { ThankYou } from './thankYou';
